requirements.txt
Pillow
gTTS
pytesseract